requirements.txt
kaggle
pandas
psycopg2-binary

